Move models and utils to common
