+++ nfdi_search_engine/common/formatting.py
