+++ nfdi_search_engine/common/models/details_settings.py
