+++ nfdi_search_engine/common/models/search_settings.py
