+++ nfdi_search_engine/services/publication_details_service.py
@@ -7,9 +7,9 @@ from typing import Any, Dict, List, Optional, Set, Tuple
 import requests
 
 from objects import Article
-from nfdi_search_engine.common.details_settings import DetailsSettings
+from nfdi_search_engine.common.models.details_settings import DetailsSettings
 from nfdi_search_engine.services.tracking_service import TrackingService
-from nfdi_search_engine.util.merge import merge_objects
+from nfdi_search_engine.common.merge import merge_objects
 
 
 class PublicationDetailsService:
